+++ packages/api/src/user.js
@@ -9,6 +9,10 @@ import {
   READ_ONLY,
   maintenanceHandler
 } from './maintenance.js'
+import { pagination } from './utils/pagination.js'
+import { toPinStatusResponse } from './pins.js'
+import { validateSearchParams } from './utils/psa.js'
+import { magicLinkBypassForE2ETestingInTestmode } from './magic.link.js'
 
 /**
  * @typedef {{ _id: string, issuer: string }} User
@@ -28,6 +32,36 @@ export async function userLoginPost (request, env) {
 }
 
 /**
+ * Controller for logging in using a magic.link token
+ * @param {Magic} magic - magic sdk instance
+ */
+function createMagicLoginController (env, testModeBypass = magicLinkBypassForE2ETestingInTestmode) {
+  const createTestmodeMetadata = (token) => {
+    const { issuer } = testModeBypass.authenticateMagicToken(env, token)
+    return {
+      issuer,
+      email: 'testMode@magic.link',
+      publicAddress: issuer
+    }
+  }
+  /**
+   * authenticate an incoming request that has a magic.link token.
+   * throws error if token isnt valid
+   * @returns {Promise} metadata about the validated token
+   */
+  const authenticate = async ({ token }) => {
+    if (testModeBypass.isEnabledForToken(env, token)) {
+      return createTestmodeMetadata(token)
+    }
+    await env.magic.token.validate(token)
+    return env.magic.users.getMetadataByToken(token)
+  }
+  return {
+    authenticate
+  }
+}
+
+/**
  * @param {Request} request
  * @param {import('./env').Env} env
  */
@@ -36,9 +70,7 @@ async function loginOrRegister (request, env) {
   const auth = request.headers.get('Authorization') || ''
 
   const token = env.magic.utils.parseAuthorizationHeader(auth)
-  env.magic.token.validate(token)
-
-  const metadata = await env.magic.users.getMetadataByToken(token)
+  const metadata = await (createMagicLoginController(env).authenticate({ token }))
   const { issuer, email, publicAddress } = metadata
   if (!issuer || !email || !publicAddress) {
     throw new Error('missing required metadata')
@@ -221,9 +253,6 @@ export async function userTokensDelete (request, env) {
   return new JSONResponse(res)
 }
 
-const sortableValues = ['Name', 'Date']
-const sortableOrders = ['Asc', 'Desc']
-
 /**
  * Retrieve a page of user uploads.
  *
@@ -234,54 +263,7 @@ export async function userUploadsGet (request, env) {
   const requestUrl = new URL(request.url)
   const { searchParams } = requestUrl
 
-  let size = 25
-  if (searchParams.has('size')) {
-    const parsedSize = parseInt(searchParams.get('size'))
-    if (isNaN(parsedSize) || parsedSize <= 0 || parsedSize > 1000) {
-      throw Object.assign(new Error('invalid page size'), { status: 400 })
-    }
-    size = parsedSize
-  }
-
-  let offset = 0
-  let page = 1
-  if (searchParams.has('page')) {
-    const parsedPage = parseInt(searchParams.get('page'))
-    if (isNaN(parsedPage) || parsedPage <= 0) {
-      throw Object.assign(new Error('invalid page number'), { status: 400 })
-    }
-    offset = (parsedPage - 1) * size
-    page = parsedPage
-  }
-
-  let before
-  if (searchParams.has('before')) {
-    const parsedBefore = new Date(searchParams.get('before'))
-    if (isNaN(parsedBefore.getTime())) {
-      throw Object.assign(new Error('invalid before date'), { status: 400 })
-    }
-    before = parsedBefore.toISOString()
-  }
-
-  let after
-  if (searchParams.has('after')) {
-    const parsedAfter = new Date(searchParams.get('after'))
-    if (isNaN(parsedAfter.getTime())) {
-      throw Object.assign(new Error('invalid after date'), { status: 400 })
-    }
-    after = parsedAfter.toISOString()
-  }
-
-  const sortBy = searchParams.get('sortBy') || 'Date'
-  const sortOrder = searchParams.get('sortOrder') || 'Desc'
-
-  if (!sortableOrders.includes(sortOrder)) {
-    throw Object.assign(new Error(`Sort ordering by '${sortOrder}' is not supported. Supported sort orders are: [${sortableOrders.toString()}]`), { status: 400 })
-  }
-
-  if (!sortableValues.includes(sortBy)) {
-    throw Object.assign(new Error(`Sorting by '${sortBy}' is not supported. Supported sort orders are: [${sortableValues.toString()}]`), { status: 400 })
-  }
+  const { size, page, offset, before, after, sortBy, sortOrder } = pagination(searchParams)
 
   const data = await env.db.listUploads(request.auth.user._id, {
     size,
@@ -348,6 +330,72 @@ export async function userUploadsRename (request, env) {
 }
 
 /**
+ * List a user's pins regardless of the token used.
+ * As we don't want to scope the Pinning Service API to users
+ * we need a new endpoint as an umbrella.
+ *
+ * @param {AuthenticatedRequest} request
+ * @param {import('./env').Env} env
+ */
+export async function userPinsGet (request, env) {
+  const requestUrl = new URL(request.url)
+  const { searchParams } = requestUrl
+
+  const { size, page, offset, before, after, sortBy, sortOrder } = pagination(searchParams)
+  const urlParams = new URLSearchParams(requestUrl.search)
+  const params = Object.fromEntries(urlParams)
+
+  const psaParams = validateSearchParams(params)
+  if (psaParams.error) {
+    throw psaParams.error
+  }
+
+  const tokens = (await env.db.listKeys(request.auth.user._id)).map((key) => key._id)
+
+  let pinRequests
+
+  try {
+    pinRequests = await env.db.listPsaPinRequests(tokens, {
+      ...psaParams.data,
+      limit: size,
+      offset,
+      before,
+      after,
+      sortBy,
+      sortOrder
+    })
+  } catch (e) {
+    console.error(e)
+    throw new HTTPError('No pinning resources found for user', 404)
+  }
+
+  const pins = pinRequests.results.map((pinRequest) => toPinStatusResponse(pinRequest))
+
+  let link = ''
+  // If there's more results to show...
+  if (page > 1) {
+    link += `<${requestUrl.pathname}?size=${size}&page=${page - 1}>; rel="previous"`
+  }
+
+  if (pins.length + offset < pinRequests.count) {
+    if (link !== '') link += ', '
+    link += `<${requestUrl.pathname}?size=${size}&page=${page + 1}>; rel="next"`
+  }
+
+  const headers = {
+    Count: pinRequests.count || 0,
+    Size: size,
+    Page: page,
+    Link: link
+  }
+
+  return new JSONResponse({
+    count: pinRequests.count,
+    results: pins
+  }, { headers })
+}
+
+/**
  *
  * @param {number} userId
  * @param {string} userProposalForm
@@ -377,7 +425,7 @@ const notifySlack = async (
     return
   }
 
-  fetch(webhookUrl, {
+  globalThis.fetch(webhookUrl, {
     method: 'POST',
     headers: {
       'Content-type': 'application/json'
@@ -408,5 +456,47 @@ const notifySlack = async (
         .join('')}
 `
     })
+  })
+}
+
+/**
+ * Get a user's payment settings.
+ *
+ * @param {AuthenticatedRequest} request
+ * @param {import('./env').Env} env
+ */
+export async function userPaymentGet (request, env) {
+  const { searchParams } = new URL(request.url)
+  const paramMethodId = searchParams.get('method.id')
+  const userPaymentGetResponse = {
+    method: paramMethodId ? { id: paramMethodId } : null
+  }
+  return new JSONResponse(userPaymentGetResponse)
+}
+
+/**
+ * Save a user's payment settings.
+ *
+ * @param {AuthenticatedRequest} request
+ * @param {import('./env').Env} env
+ */
+export async function userPaymentPut (request, env) {
+  const requestBody = await request.json()
+  const method = requestBody?.method?.id
+    ? { id: requestBody?.method?.id }
+    : {
+      // stubbing this for now with the value from the docs.
+      //   https://stripe.com/docs/api/payment_methods/object
+        id: env.mockStripePaymentMethodId,
+        warning: 'this method is a stub. The id here is different than anything you might have sent in the request.'
+      }
+  const savePaymentSettingsResponse = {
+    method
+  }
+  return new JSONResponse(savePaymentSettingsResponse, {
+    status: 202,
+    headers: {
+      location: `/user/payment?method.id=${method.id}`
+    }
   })
 }